Fix:- Identify week points in clean coding and finished.

+++ Backend/controllers/product.controller.js
@@ -46,11 +46,15 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
     const { id } = req.params;
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: 'Invalid product ID' }); // invalid ID format
+    }
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: 'Product deleted successfully' });
     } catch (error) {
         console.log('Error in deleting product:', error.message);
-        res.status(404).json({ success: false, message: 'Product not found' });
+        res.status(500).json({ success: false, message: 'Server Error' });
     }
 };
